changed data type in nearest location c program

+++ cse2124_w26/lab_3/nearest_location.c
@@ -6,18 +6,18 @@ among Saidpur, Boraibari, and Kamalpur using latitude and longitude.
 
 int main() {
     // Given locations
-    float lat1 = 25.778, lon1 = 88.897;  // Saidpur
-    float lat2 = 25.513, lon2 = 89.856;  // Boraibari
-    float lat3 = 25.286, lon3 = 89.863;  // Kamalpur
+    double lat1 = 25.778, lon1 = 88.897;  // Saidpur
+    double lat2 = 25.513, lon2 = 89.856;  // Boraibari
+    double lat3 = 25.286, lon3 = 89.863;  // Kamalpur
 
     // User input location
-    float lat, lon;
-    scanf("%f %f", &lat, &lon);
+    double lat, lon;
+    scanf("%lf %lf", &lat, &lon);
 
     // Distance calculation (simple squared distance)
-    float d1 = (lat - lat1)*(lat - lat1) + (lon - lon1)*(lon - lon1);
-    float d2 = (lat - lat2)*(lat - lat2) + (lon - lon2)*(lon - lon2);
-    float d3 = (lat - lat3)*(lat - lat3) + (lon - lon3)*(lon - lon3);
+    double d1 = (lat - lat1)*(lat - lat1) + (lon - lon1)*(lon - lon1);
+    double d2 = (lat - lat2)*(lat - lat2) + (lon - lon2)*(lon - lon2);
+    double d3 = (lat - lat3)*(lat - lat3) + (lon - lon3)*(lon - lon3);
 
     // Compare using conditional statements
     if(d1 < d2 && d1 < d3) {
